+++ src/models/UnitStatus.ts
@@ -5,12 +5,22 @@ export enum UnitStates {
 
 export class UnitStatus {
   private readonly _state: UnitStates;
+  private readonly _brightness: number;
 
-  constructor(status: number) {
+  constructor(status: UnitStates, brightness?: number) {
     this._state = status;
+    if (brightness !== undefined) {
+      this._brightness = brightness;
+    } else {
+      this._brightness = this.state === UnitStates.Off ? 0 : 100;
+    }
   }
 
-  get state(): number {
+  get state(): UnitStates {
     return this._state;
+  }
+
+  get brightness(): number {
+    return this._brightness;
   }
 }