+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## NEXT
 
 * Prevent garage door open/close command if currently in opening/closing state
+* Support brightness level for Omni units
 
 ## 1.3.0 (2020-11-14)